% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateAlignments.R
\name{addProteinNames}
\alias{addProteinNames}
\title{Annotates protein names as the top GRangesList structure}
\usage{
addProteinNames(alignment, idmap)
}
\arguments{
\item{alignment}{Alignment output, A \code{GRangesList} with peptide sequences as names}

\item{idmap}{a map between peptide ID and Protein/Gene ID.
Must be a \code{data.table} object with columns \code{"id", "protein_id"}.}
}
\value{
A GRanges object with the genomic coordinates of the queried peptide
}
\description{
Annotates protein names as the top GRangesList structure
}

