% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignPeptideToGene.R
\name{alignPeptideToGene}
\alias{alignPeptideToGene}
\title{Aligns a provided peptide sequence to a provided Gene sequence.}
\usage{
alignPeptideToGene(peptide, geneID, exon, proteinseq, idmap)
}
\arguments{
\item{peptide}{Peptide sequence as character}

\item{geneID}{The geneID from which the peptide originated}

\item{exon}{A \code{data.table} with exon information, obtained by a download function(see ...)}

\item{proteinseq}{a table with protein sequences for every transcript in the reference genome.}

\item{idmap}{a map between geneID and Transcript ID. Must be a keyed \code{data.table} object.}
}
\value{
A GRanges object with the genomic coordinates of the queried peptide
}
\description{
Aligns a provided peptide sequence to a provided Gene sequence.
}
\examples{
alignPeptideToGene(peptide = "LHAVVETLVNHR", geneID = "ENSG00000028528",
                   exon = exon_dt, proteinseq, idmap)
}

