% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignPeptideToGene.R
\name{alignPeptideToTranscript}
\alias{alignPeptideToTranscript}
\title{Aligns a provided peptide sequence to a provided Transcript.}
\usage{
alignPeptideToTranscript(peptide, transcriptID, exon, proteinseq)
}
\arguments{
\item{peptide}{Peptide sequence as character}

\item{transcriptID}{The transcriptID from which the peptide possibly originated}

\item{exon}{A \code{data.table} with exon information, obtained by a download function(see ...)}

\item{proteinseq}{a table with protein sequences for every transcript in the reference genome.}
}
\value{
A GRanges object with the genomic coordinates of the queried peptide
}
\description{
Aligns a provided peptide sequence to a provided Transcript.
}

