% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignPepTraces.R
\name{alignPepTraces}
\alias{alignPepTraces}
\title{Aligns peptides in a pepTraces object to a reference genome}
\usage{
alignPepTraces(pepTraces, exon, proteinseq, mapIDs = FALSE,
  mappingTable = NULL)
}
\value{
A GRangesList object with the genomic position of the found peptides
}
\description{
Aligns peptides in a pepTraces object to a reference genome
}

